+++ app/policies/project_policy.rb
@@ -46,10 +46,6 @@ class ProjectPolicy
     project_assigned_to_user?
   end
 
-  def openers?
-    true
-  end
-
   def change_conversion_date?
     return false if @record.conversion_date_provisional?